Fix null log

+++ GeneracionNumeracionAvalonia/Services/ILogger.cs
@@ -19,7 +19,11 @@ namespace GeneracionNumeracionAvalonia.Services
             string logFileName = "GeneradorNumeracion.log";
 
             string logFilePath = Path.Combine(documentsPath, appName, logFileName);
-            File.Delete(logFilePath);
+            if (string.IsNullOrEmpty(logFilePath))
+            {
+                File.Delete(logFilePath);
+            }
+
 
             Log.Logger = new LoggerConfiguration()
                 .MinimumLevel.Debug()
